Deleted duplicate model, clean up of classes, moved interfaces to own file
