--- .github/workflows/Old invoice_gen.sln notepad contents.txt
@@ -1,68 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{C602692F-1C12-4715-AF61-CAE40E581C8D}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{834C3722-1EF2-45F0-83C3-680EC44BA8E0}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{15D34F74-5116-4BA8-8380-C38572BE2652}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Invoice_Gen.WebApi", "src\Invoice_Gen.WebApi\Invoice_Gen.WebApi.csproj", "{A744E8D9-2EEA-4A89-8141-C444E862129C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Invoice_Gen.WebApi.IntegrationTests", "tests\Invoice_Gen.WebApi.IntegrationTests\Invoice_Gen.WebApi.IntegrationTests.csproj", "{09B0E22F-CDBC-4607-BE14-881FE56BEB9B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Invoice_GenUI", "src\Invoice_GenUI\Invoice_GenUI\Invoice_GenUI.csproj", "{7F291313-167E-426E-AB17-C07DC4E5B0A2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Invoice_Gen.ViewModels", "src\Invoice_Gen.ViewModels\Invoice_Gen.ViewModels.csproj", "{76E98319-0968-4CDA-885F-C180175C6AA3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Invoice_Gen.WebApi.UnitTests", "tests\Invoice_Gen.WebApi.UnitTests\Invoice_Gen.WebApi.UnitTests.csproj", "{E6FC56C0-95BE-4EA2-9002-89D9954E6B09}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "devops", "devops", "{C02A8D52-FC59-43A2-AC67-49117705277B}"
-	ProjectSection(SolutionItems) = preProject
-		.github\workflows\main_2023-invoice-gen.yml = .github\workflows\main_2023-invoice-gen.yml
-		.github\workflows\webapi-pr-action.yml = .github\workflows\webapi-pr-action.yml
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{A744E8D9-2EEA-4A89-8141-C444E862129C} = {834C3722-1EF2-45F0-83C3-680EC44BA8E0}
-		{09B0E22F-CDBC-4607-BE14-881FE56BEB9B} = {15D34F74-5116-4BA8-8380-C38572BE2652}
-		{7F291313-167E-426E-AB17-C07DC4E5B0A2} = {834C3722-1EF2-45F0-83C3-680EC44BA8E0}
-		{76E98319-0968-4CDA-885F-C180175C6AA3} = {834C3722-1EF2-45F0-83C3-680EC44BA8E0}
-		{E6FC56C0-95BE-4EA2-9002-89D9954E6B09} = {15D34F74-5116-4BA8-8380-C38572BE2652}
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A744E8D9-2EEA-4A89-8141-C444E862129C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A744E8D9-2EEA-4A89-8141-C444E862129C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A744E8D9-2EEA-4A89-8141-C444E862129C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A744E8D9-2EEA-4A89-8141-C444E862129C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{09B0E22F-CDBC-4607-BE14-881FE56BEB9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{09B0E22F-CDBC-4607-BE14-881FE56BEB9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{09B0E22F-CDBC-4607-BE14-881FE56BEB9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{09B0E22F-CDBC-4607-BE14-881FE56BEB9B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7F291313-167E-426E-AB17-C07DC4E5B0A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7F291313-167E-426E-AB17-C07DC4E5B0A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7F291313-167E-426E-AB17-C07DC4E5B0A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7F291313-167E-426E-AB17-C07DC4E5B0A2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{76E98319-0968-4CDA-885F-C180175C6AA3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{76E98319-0968-4CDA-885F-C180175C6AA3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{76E98319-0968-4CDA-885F-C180175C6AA3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{76E98319-0968-4CDA-885F-C180175C6AA3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E6FC56C0-95BE-4EA2-9002-89D9954E6B09}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E6FC56C0-95BE-4EA2-9002-89D9954E6B09}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E6FC56C0-95BE-4EA2-9002-89D9954E6B09}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E6FC56C0-95BE-4EA2-9002-89D9954E6B09}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
