+++ src/Invoice_GenUI/Invoice_GenUI/Models/Services/INavigationService.cs
@@ -1,0 +1,9 @@
+﻿namespace Invoice_GenUI.Models.Services
+{
+    public interface INavigationService
+    {
+        ViewModel CurrentView { get; }
+        void NavigateTo<T>() where T : ViewModel;
+        void ParameterNavigateTo<T>(object parameter) where T : ViewModel;
+    }
+}
