+++ src/Invoice_GenUI/Invoice_GenUI/Models/Services/NavigationService.cs
@@ -2,12 +2,6 @@
 
 namespace Invoice_GenUI.Models.Services
 {
-    public interface INavigationService
-    {
-        ViewModel CurrentView { get; }
-        void NavigateTo<T>() where T : ViewModel;
-        void ParameterNavigateTo<T>(object parameter) where T : ViewModel;
-    }
     public class NavigationService : ViewModel, INavigationService
     {
         private readonly Func<Type, ViewModel> _viewModelFactory;
