--- src/Invoice_GenUI/Invoice_GenUI/Models/SingleClientModel.cs
@@ -1,11 +1,0 @@
-﻿namespace Invoice_GenUI.Models
-{
-    public class SingleClientModel
-    {
-        public int ClientId { get; set; }
-        public string? ClientName { get; set; }
-        public string? ClientAddress { get; set; }
-        public string? ContactEmail { get; set; }
-        public string? ContactName { get; set; }
-    }
-}
